+++ web/package.json
@@ -18,7 +18,7 @@
   "dependencies": {
     "@axa-ch/react-polymorphic-types": "^1.4.2",
     "@floating-ui/react": "^0.27.19",
-    "@inlang/paraglide-js": "^2.18.2",
+    "@inlang/paraglide-js": "^2.19.0",
     "@react-hook/resize-observer": "^2.0.2",
     "@shortercode/webzip": "1.1.1-0",
     "@stablelib/base64": "^2.0.1",
@@ -54,7 +54,7 @@
     "zustand": "^5.0.14"
   },
   "devDependencies": {
-    "@biomejs/biome": "2.4.15",
+    "@biomejs/biome": "2.4.16",
     "@inlang/paraglide-js": "2.18.1",
     "@tanstack/devtools-vite": "^0.7.0",
     "@tanstack/react-devtools": "^0.10.5",
@@ -64,7 +64,7 @@
     "@types/byte-size": "^8.1.2",
     "@types/humanize-duration": "^3.27.4",
     "@types/lodash-es": "^4.17.12",
-    "@types/node": "^25.9.2",
+    "@types/node": "^25.9.3",
     "@types/qs": "^6.15.1",
     "@types/react": "^19.2.17",
     "@types/react-dom": "^19.2.3",
@@ -74,7 +74,7 @@
     "globals": "^17.6.0",
     "prettier": "^3.8.4",
     "sass": "^1.100.0",
-    "sharp": "^0.34.5",
+    "sharp": "^0.35.0",
     "stylelint": "^17.13.0",
     "stylelint-config-standard-scss": "^17.0.0",
     "stylelint-scss": "^7.2.0",
