+++ web/src/pages/settings/SettingsSmtpPage/SettingsSmtpPage.tsx
@@ -1,3 +1,4 @@
+import './style.scss';
 import { useMutation, useQuery } from '@tanstack/react-query';
 import { Link } from '@tanstack/react-router';
 import { useMemo, useRef, useState } from 'react';
